+++ app/sprinkles/account/src/Controller/AccountController.php
@@ -361,10 +361,9 @@ class AccountController extends SimpleController
         $ms->addMessageTranslated('success', 'WELCOME', $currentUser->export());
 
         // Set redirect, if relevant
-        $determineRedirectOnLogin = $this->ci->determineRedirectOnLogin;
-        $response = $determineRedirectOnLogin($response);
+        $redirectOnLogin = $this->ci->get('redirect.onLogin');
 
-        return $response->withStatus(200);
+        return $redirectOnLogin($request, $response, $args);
     }
 
     /**
@@ -424,10 +423,11 @@ class AccountController extends SimpleController
         /** @var UserFrosting\Sprinkle\Account\Authenticate\Authenticator $authenticator */
         $authenticator = $this->ci->authenticator;
 
-        // Forward to dashboard if user is already logged in
-        // TODO: forward to user's landing page or last visited page
+        // Redirect if user is already logged in
         if ($authenticator->check()) {
-            return $response->withRedirect($this->ci->router->pathFor('dashboard'), 302);
+            $redirect = $this->ci->get('redirect.onAlreadyLoggedIn');
+
+            return $redirect($request, $response, $args);
         }
 
         // Load validation rules
@@ -577,10 +577,11 @@ class AccountController extends SimpleController
         /** @var UserFrosting\Sprinkle\Account\Authenticate\Authenticator $authenticator */
         $authenticator = $this->ci->authenticator;
 
-        // Forward to dashboard if user is already logged in
-        // TODO: forward to user's landing page or last visited page
+        // Redirect if user is already logged in
         if ($authenticator->check()) {
-            return $response->withRedirect($this->ci->router->pathFor('dashboard'), 302);
+            $redirect = $this->ci->get('redirect.onAlreadyLoggedIn');
+
+            return $redirect($request, $response, $args);
         }
 
         // Load validation rules
@@ -814,7 +815,7 @@ class AccountController extends SimpleController
         Capsule::transaction( function() use ($classMapper, $data, $ms, $config, $throttler) {
             // Log throttleable event
             $throttler->logEvent('registration_attempt');
-        
+
             // Create the user
             $user = $classMapper->createInstance('user', $data);
 
@@ -1003,14 +1004,10 @@ class AccountController extends SimpleController
 
         $ms->addMessageTranslated('success', 'PASSWORD.UPDATED');
 
-        // Log out any existing user, and create a new session
-
-        /** @var UserFrosting\Sprinkle\Account\Database\Models\User $currentUser */
-        $currentUser = $this->ci->currentUser;
-
         /** @var UserFrosting\Sprinkle\Account\Authenticate\Authenticator $authenticator */
         $authenticator = $this->ci->authenticator;
 
+        // Log out any existing user, and create a new session
         if ($authenticator->check()) {
             $authenticator->logout();
         }