+++ app/sprinkles/admin/src/ServicesProvider/ServicesProvider.php
@@ -7,6 +7,8 @@
  */
 namespace UserFrosting\Sprinkle\Admin\ServicesProvider;
 
+use Psr\Http\Message\ResponseInterface as Response;
+use Psr\Http\Message\ServerRequestInterface as Request;
 use UserFrosting\Sprinkle\Account\Authenticate\Authenticator;
 use UserFrosting\Sprinkle\Account\Authorize\AuthorizationManager;
 use UserFrosting\Sprinkle\Core\Facades\Debug;
@@ -43,10 +45,27 @@ class ServicesProvider
 
         /**
          * Returns a callback that handles setting the `UF-Redirect` header after a successful login.
+         *
+         * Overrides the service definition in the account Sprinkle.
          */
-        $container['determineRedirectOnLogin'] = function ($c) {
-            return function ($response) use ($c)
-            {
+        $container['redirect.onLogin'] = function ($c) {
+            /**
+             * This method is invoked when a user completes the login process.
+             *
+             * Returns a callback that handles setting the `UF-Redirect` header after a successful login.
+             * @param \Psr\Http\Message\ServerRequestInterface $request  
+             * @param \Psr\Http\Message\ResponseInterface      $response 
+             * @param array $args
+             * @return \Psr\Http\Message\ResponseInterface
+             */
+            return function (Request $request, Response $response, array $args) use ($c) {
+                // Backwards compatibility for the deprecated determineRedirectOnLogin service
+                if ($c->has('determineRedirectOnLogin')) {
+                    $determineRedirectOnLogin = $c->determineRedirectOnLogin;
+            
+                    return $determineRedirectOnLogin($response)->withStatus(200);
+                }
+
                 /** @var UserFrosting\Sprinkle\Account\Authorize\AuthorizationManager */
                 $authorizer = $c->authorizer;